Fix a bug in EDU Typing handling in syncapi.
When a user's typing status times out, it must advance both the
Notifier and the EDUCache. This inconsistency meant that some Sytests
that only does a single `/sync` saw a returned event stream it didn't
expect (though empty), and failed.

In production, this likely just has the effect of adding extra load on
`/sync`, as the stream positions will resync the next time a user is
added or removed via API.