Adds an explicit empty pushrules.Rule.Conditions for all default override rules.
https://github.com/matrix-org/matrix-react-sdk/blob/fe0a68b71ea81230bb662be0c8b89a56c97542bc/src/RoomNotifs.ts#L226
crashes if "override" rules don't all have condition arrays. JSON
encoding in Go is quirky in that "nil" becomes "null" rather than
"[]". This also means we cannot use "omitempty" because that omits
both the "nil" and "[]" cases.

I consider the SDK wrong here, since conditions is not marked as
required in the (unstable) spec. But that code is 5 years old, so we'd
better comply.